move main.go to cmd/cli
