+++ Makefile
@@ -13,18 +13,16 @@ GOLANG_CROSS_VERSION  ?= v1.19.0
 .PHONY: run build release install lint format test integration tidy clean release-windows release-macos release-linux
 
 run: build
-	go run .
+	go run ./cmd/cli/
 
 build:
 	@mkdir -p ${BINARY_DIR}
 ifeq ($(OS),Windows_NT)
 	@go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-cli.exe ./cmd/cli && \
-	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher ./ && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-marker.exe ./cmd/marker && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-github.exe ./cmd/github-action
 else
 	@go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-cli ./cmd/cli && \
-	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher ./ && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-marker ./cmd/marker && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-github-action ./cmd/github-action
 endif
@@ -33,12 +31,10 @@ release:
 	@mkdir -p ${BINARY_DIR}
 ifeq ($(OS),Windows_NT)
 	@go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-cli.exe ./cmd/cli && \
-	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher ./ && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-marker.exe ./cmd/marker && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-github.exe ./cmd/github-action
 else
 	@go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-cli ./cmd/cli && \
-	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher ./ && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-marker ./cmd/marker && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-github-action ./cmd/github-action
 endif
@@ -49,7 +45,6 @@ release-windows:
 	export CGO_ENABLED=1 && \
 	echo Building $$GOOS-$$GOARCH && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-cli-$$GOOS-$$GOARCH.exe ./cmd/cli && \
-	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-$$GOOS-$$GOARCH.exe ./ && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-marker-$$GOOS-$$GOARCH.exe ./cmd/marker && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-github-action-$$GOOS-$$GOARCH.exe ./cmd/github-action
 
@@ -59,7 +54,6 @@ release-macos:
 	export CGO_ENABLED=1 && \
 	echo Building $$GOOS-$$GOARCH && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-cli-$$GOOS-$$GOARCH.exe ./cmd/cli && \
-	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-$$GOOS-$$GOARCH.exe ./ && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-marker-$$GOOS-$$GOARCH.exe ./cmd/marker && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-github-action-$$GOOS-$$GOARCH.exe ./cmd/github-action
 
@@ -70,7 +64,6 @@ release-linux:
 	export CGO_ENABLED=1 && \
 	echo Building $$GOOS-$$GOARCH && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-cli-$$GOOS-$$GOARCH.exe ./cmd/cli && \
-	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-$$GOOS-$$GOARCH.exe ./ && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-marker-$$GOOS-$$GOARCH.exe ./cmd/marker && \
 	go build ${LDFLAGS_RELEASE} -o ${BINARY_DIR}/go-gopher-github-action-$$GOOS-$$GOARCH.exe ./cmd/github-action
