+++ cmd/github-action/commands.go
@@ -88,9 +88,9 @@ func actionCommand(cCtx *cli.Context) error {
 		violations = violation.FilterByLogin(violations, whitelist)
 	}
 
-	workflowSummary(authors, violated, count, total, violations)
+	workflow.PrintSummary(authors, violated, count, total, violations)
 
-	summary := markdownSummary(authors, violations)
+	summary := workflow.MarkdownSummary(authors, violations)
 	markup.Outputs("pr_summary", summary)
 
 	fn, err := ghwf.WriteLog(*enrichedModel, cfg)
@@ -107,150 +107,4 @@ func actionCommand(cCtx *cli.Context) error {
 	}
 
 	return nil
-}
-
-// Print violation summary to IO, Split by severity with author association.
-func workflowSummary(authors utils.Authors, v, c, t int, vs []violation.Violation) {
-	var violations, suggestions []violation.Violation
-	for _, v := range vs {
-		switch v.Severity() {
-		case violation.Violated:
-			violations = append(violations, v)
-		case violation.Suggestion:
-			suggestions = append(suggestions, v)
-		}
-	}
-
-	var vSd strings.Builder
-	for _, v := range violations {
-		vSd.WriteString(v.Display(authors))
-	}
-	markup.Group("Violations", vSd.String())
-
-	var sSd strings.Builder
-	for _, v := range suggestions {
-		sSd.WriteString(v.Display(authors))
-	}
-	markup.Group("Suggestions", sSd.String())
-
-	var aSd strings.Builder
-	counts := make(map[string]int)
-	for _, v := range vs {
-		email := v.Email()
-		login, err := authors.Find(email)
-		if err != nil {
-			continue
-		}
-		counts[*login]++
-	}
-
-	for login, count := range counts {
-		aSd.WriteString(fmt.Sprintf("%s: %d\n", login, count))
-	}
-
-	aSd.WriteString(fmt.Sprintf("violated: %d\n", v))
-	aSd.WriteString(fmt.Sprintf("count: %d\n", c))
-	aSd.WriteString(fmt.Sprintf("total: %d\n", t))
-	markup.Group("Summary", aSd.String())
-}
-
-// Helper function to create a markdown summary of the violations.
-// nolint: gocognit
-func markdownSummary(authors utils.Authors, vs []violation.Violation) string {
-	md := markup.CreateMarkdown("Workflow Summary")
-	md.AddLine(fmt.Sprintf("Created with git-gopher version `%s`", version.BuildVersion()))
-
-	// Separate violation types.
-	var violations []violation.Violation
-	var suggestions []violation.Violation
-
-	for _, v := range vs {
-		switch v.Severity() {
-		case violation.Violated:
-			if v.Current() {
-				violations = append(violations, v)
-			}
-		case violation.Suggestion:
-			if v.Current() {
-				suggestions = append(suggestions, v)
-			}
-		default:
-			log.Printf("Unknown violation severity: %v", v.Severity())
-		}
-	}
-
-	if len(violations) > 0 {
-		headers := []string{"Violation", "Message", "Advice", "Author"}
-		rows := make([][]string, len(violations))
-
-		for i, v := range violations {
-			row := make([]string, len(headers))
-			name := v.Name()
-			row[0] = name
-			message := v.Message()
-			row[1] = message
-
-			suggestion, err := v.Suggestion()
-			if err != nil {
-				suggestion = ""
-			}
-			row[2] = suggestion
-
-			usernamePtr, err := authors.Find(v.Email())
-			if err != nil || usernamePtr == nil {
-				row[3] = "unknown"
-			} else {
-				row[3] = markup.Author(*usernamePtr).Markdown()
-			}
-
-			rows[i] = row
-		}
-
-		md.BeginCollapsable("Violations")
-		md.Table(headers, rows)
-		md.EndCollapsable()
-	}
-
-	if len(suggestions) > 0 {
-		headers := []string{"Suggestion", "Message", "Advice", "Author"}
-		rows := make([][]string, len(suggestions))
-
-		for i, v := range suggestions {
-			row := make([]string, len(headers))
-			name := v.Name()
-			row[0] = name
-			message := v.Message()
-			row[1] = message
-
-			suggestion, err := v.Suggestion()
-			if err != nil {
-				suggestion = ""
-			}
-			row[2] = suggestion
-
-			usernamePtr, err := authors.Find(v.Email())
-			if err != nil || usernamePtr == nil {
-				row[3] = "unknown"
-			} else {
-				row[3] = markup.Author(*usernamePtr).Markdown()
-			}
-
-			rows[i] = row
-		}
-
-		md.BeginCollapsable("Suggestions")
-		md.Table(headers, rows)
-		md.EndCollapsable()
-	}
-
-	workflowUrl := os.Getenv("WORKFLOW_URL")
-	if (len(violations)+len(suggestions)) < len(vs) && workflowUrl != "" {
-		md.AddLine(fmt.Sprintf(`There still exist some violations beyond the scope of this pull request, 
-			please view the full log [here](%s)`, workflowUrl))
-	}
-
-	// Google form
-	md.AddLine(fmt.Sprintf("Have any feedback? Feel free to submit it [here](%s)", utils.GoogleFormURL))
-
-	return md.Render()
 }
